requirements.txt
Flask
praat-parselmouth
numpy
scipy
